next phase

+++ src/services/apiCabins.js
@@ -5,7 +5,6 @@ export async function getCabins() {
 
   if (error) {
     console.error(error);
-
     throw new Error("Cabins could not be loaded");
   }
 
@@ -23,29 +22,30 @@ export async function createEditCabin(newCabin, id) {
     ? newCabin.image
     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // 1. Create edit/cabin
+  // 1. Create/edit cabin
   let query = supabase.from("cabins");
 
-  // a) Create
+  // A) CREATE
   if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-  // b) Edit
+  // B) EDIT
   if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
 
   const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
-    throw new Error("Cabin could not be deleted");
+    throw new Error("Cabin could not be created");
   }
 
   // 2. Upload image
+  if (hasImagePath) return data;
 
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
-  // 3. Delete the cabin IF there was an error uploading image
 
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
@@ -53,6 +53,7 @@ export async function createEditCabin(newCabin, id) {
       "Cabin image could not be uploaded and the cabin was not created"
     );
   }
+
   return data;
 }
